Add files via upload
This dataset aimed at the case of customers' default payments in Taiwan and compares the predictive accuracy of probability of default category. From the perspective of risk management, the result of predictive accuracy of the estimated probability of default will be more valuable than the binary result of classification - credible or not credible clients.