Add files via upload
Business Scenario: Say there are 5 marketing offers that need to be mapped to individual customers. The business wants you to do a quick analysis to figure out which offer should be given to which customer. Potential Solution: Perform PCA and then build clusters on the Components or Build clusters on the given data after scaling. When you profile the clusters, the behaviors are likely to differ across clusters. Based on the behavior, you can decide which offer should be given to which group.